sort/hilbert_kernel: fix invalid symbol type.
This bug did not cause a compiler error because up to now device
constants have been passed to cuda::symbol by const reference.
cuda::symbol will be changed to require a non-const reference for
initialisation in an upcoming commit.

+++ halmd/mdsim/gpu/sort/hilbert_kernel.hpp
@@ -34,7 +34,7 @@ struct hilbert_wrapper
     typedef typename type_traits<dimension, float>::gpu::vector_type vector_type;
 
     /** Hilbert space-filling curve recursion depth */
-    cuda::symbol<float> depth;
+    cuda::symbol<unsigned int> depth;
     /** cubic box edgle length */
     cuda::symbol<vector_type> box_length;
     /** generate Hilbert space-filling curve */
